large refactor, broken
